requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
oracledb==2.5.1
httpx==0.28.1
python-multipart==0.0.20
jinja2==3.1.5
aiofiles==24.1.0

# OpenTelemetry
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-instrumentation-sqlalchemy==0.50b0
opentelemetry-instrumentation-httpx==0.50b0
opentelemetry-instrumentation-logging==0.50b0

# OCI SDK (for logging)
oci==2.143.0

# Security (bcrypt for password hashing)
bcrypt==4.2.1
